Revert "[CSB-3069] - Opensearch system-x implementation"
This reverts commit 039d919221198882278b6249e71a1b2cb10dc65b.
