+++ bom/pom.xml
@@ -171,16 +171,6 @@
             </dependency>
             <dependency>
                 <groupId>software.tnb</groupId>
-                <artifactId>system-x-search-engine</artifactId>
-                <version>1.0-SNAPSHOT</version>
-            </dependency>
-            <dependency>
-                <groupId>software.tnb</groupId>
-                <artifactId>system-x-search-common</artifactId>
-                <version>1.0-SNAPSHOT</version>
-            </dependency>
-            <dependency>
-                <groupId>software.tnb</groupId>
                 <artifactId>system-x-elasticsearch</artifactId>
                 <version>1.0-SNAPSHOT</version>
             </dependency>
@@ -377,11 +367,6 @@
             <dependency>
                 <groupId>software.tnb</groupId>
                 <artifactId>system-x-mysql</artifactId>
-                <version>1.0-SNAPSHOT</version>
-            </dependency>
-            <dependency>
-                <groupId>software.tnb</groupId>
-                <artifactId>system-x-opensearch</artifactId>
                 <version>1.0-SNAPSHOT</version>
             </dependency>
             <dependency>
