+++ system-x/services/all/pom.xml
@@ -123,10 +123,6 @@
         </dependency>
         <dependency>
             <groupId>software.tnb</groupId>
-            <artifactId>system-x-search-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>software.tnb</groupId>
             <artifactId>system-x-elasticsearch</artifactId>
         </dependency>
         <dependency>
@@ -268,10 +264,6 @@
         <dependency>
             <groupId>software.tnb</groupId>
             <artifactId>system-x-mysql</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>software.tnb</groupId>
-            <artifactId>system-x-opensearch</artifactId>
         </dependency>
         <dependency>
             <groupId>software.tnb</groupId>
