+++ system-x/services/pom.xml
@@ -36,7 +36,7 @@
         <module>infinispan</module>
         <module>db</module>
         <module>mail</module>
-        <module>search-engine</module>
+        <module>elasticsearch</module>
         <module>filesystem</module>
         <module>rest</module>
         <module>salesforce</module>
