--- system-x/services/search-engine/elasticsearch/pom.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>system-x-search-engine</artifactId>
-        <groupId>software.tnb</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>system-x-elasticsearch</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <name>TNB :: System-X :: Services :: ElasticSearch</name>
-
-    <properties>
-        <elasticsearch.client.version>8.15.0</elasticsearch.client.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>software.tnb</groupId>
-            <artifactId>system-x-search-common</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>testcontainers</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>co.elastic.clients</groupId>
-            <artifactId>elasticsearch-java</artifactId>
-            <version>${elasticsearch.client.version}</version>
-        </dependency>
-    </dependencies>
-</project>
