--- system-x/services/search-engine/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>software.tnb</groupId>
-        <artifactId>system-x-services</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>system-x-search-engine</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <name>TNB :: System-X :: Services :: Search engine</name>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>common</module>
-        <module>elasticsearch</module>
-        <module>opensearch</module>
-    </modules>
-
-</project>
